Updating to latest FlutterFlow output.

+++ firebase/functions/index.js
@@ -6,4 +6,16 @@ exports.onUserDeleted = functions.auth.user().onDelete(async (user) => {
   let firestore = admin.firestore();
   let userRef = firestore.doc("users/" + user.uid);
   await firestore.collection("users").doc(user.uid).delete();
+  await firestore
+    .collection("biblioteca_pictogramas")
+    .where("dono_uid", "==", user.uid)
+    .get()
+    .then(async (querySnapshot) => {
+      for (var doc of querySnapshot.docs) {
+        console.log(
+          `Deleting document ${doc.id} from collection biblioteca_pictogramas`,
+        );
+        await doc.ref.delete();
+      }
+    });
 });
